requirements.txt
altair==5.5.0
attrs==25.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.3.1
cssselect==1.3.0
cycler==0.12.1
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.18.0
fonttools==4.56.0
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
jieba3k==0.35.1
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
lxml==5.3.1
lxml_html_clean==0.1.1
MarkupSafe==3.0.2
matplotlib==3.10.1
narwhals==1.31.0
newspaper3k==0.2.8
nltk==3.9.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.4
pyarrow==19.0.1
pydeck==0.9.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
rpds-py==0.23.1
seaborn==0.13.2
streamlit == 1.43.2
textblob==0.19.0
wordcloud==1.9.4